python-projects/instagram-followers-scraper/requirements.txt
python-decouple
selenium
webdriver_manager
python-dotenv